+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 ### Added
 - [CSE] Added internal check before executing a DELETE operation that all resources are allowed to be deleted.
 - [TUI] Added auto-refreshing diagrams and checkbox to enable/disable auto-refreshing.
+- [TUI] Added validation: *notificationURI* must only be present in non-blocking async requests.
 
 ### Experimental
 
@@ -23,6 +24,8 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 ### Fixed
 - [CSE] Added missing &lt;CIN> DELETE procedure: Delete is forbidden if the parent &lt;CNT> has the *disableRetrieval* attribute set to True.
 - [CSE] Fixed wrong http response for some cases when the *Accept* http header contains a wildcard.
+- [CSE] Fixed wrong http request when forwarding non-blocking requests.
+- [CSE] Fixed empty result content in the &lt;REQ> instance when forwarding non-blocking requests to a remote CSE.
 
 ### Removed